+++ power-libperfmgr/Android.bp
@@ -19,12 +19,27 @@ cc_library_headers {
     export_include_dirs: ["."],
 }
 
+cc_library {
+    name: "libdisppower",
+    proprietary: true,
+    srcs: [
+        "disp-power/InteractionHandler.cpp",
+    ],
+     shared_libs: [
+        "libbase",
+        "libcutils",
+        "liblog",
+        "libperfmgr",
+        "libutils",
+    ],
+}
+
 cc_binary {
     name: "android.hardware.power@1.3-service.xiaomi_sm6250-libperfmgr",
     relative_install_path: "hw",
-    vintf_fragments: ["android.hardware.power@1.3-service.xiaomi_sm6250.xml"],
-    init_rc: ["android.hardware.power@1.3-service.xiaomi_sm6250-libperfmgr.rc"],
-    srcs: ["service.cpp", "Power.cpp", "InteractionHandler.cpp"],
+    vintf_fragments: ["hidl/android.hardware.power@1.3-service.xiaomi_sm6250.xml"],
+    init_rc: ["hidl/android.hardware.power@1.3-service.xiaomi_sm6250-libperfmgr.rc"],
+    srcs: ["hidl/service.cpp", "hidl/Power.cpp"],
     cflags: [
         "-Wall",
         "-Werror",
@@ -39,6 +54,7 @@ cc_binary {
         "android.hardware.power@1.1",
         "android.hardware.power@1.2",
         "android.hardware.power@1.3",
+        "libdisppower",
         "libperfmgr",
     ],
 	header_libs: [
