+++ power-libperfmgr/disp-power/InteractionHandler.cpp
