+++ power-libperfmgr/hidl/Power.h
@@ -26,7 +26,7 @@
 #include <hidl/Status.h>
 #include <perfmgr/HintManager.h>
 
-#include "InteractionHandler.h"
+#include "disp-power/InteractionHandler.h"
 
 namespace android {
 namespace hardware {
