+++ power-libperfmgr/hidl/android.hardware.power@1.3-service.xiaomi_sm6250.xml
